requirements.txt
numpy
pyopengl
pyopengl_accelerate
pysdl2